refactor(bot): extract bot/audit.py + bot/rate_limit.py + bot/path_validation.py — closes #114
Second slice of the bot/core decomposition (PRD #113, ADR 0001).
Mechanical extraction of the three smallest leaf modules; no new
behavior, no new tests. The existing test suite serves as the
regression net.

Cross-module concerns handled:

* ``bot.path_validation`` calls ``audit()`` from ``bot.audit`` directly
  — clean dep direction (audit is a leaf with no bot imports).
* ``bot.audit`` lazily resolves the audit log path on first call (via a
  late ``from bot.core import OMICSCLAW_DIR``) so it can be loaded
  while ``bot.core`` is still mid-initialisation.
* ``bot.rate_limit._evict_lru_conversations`` reaches into bot.core's
  ``transcript_store`` / ``tool_result_store`` / ``MAX_CONVERSATIONS``;
  late-imports them inside the function for the same reason.
* ``bot.path_validation`` uses ``OMICSCLAW_DIR`` / ``DATA_DIR`` /
  ``EXAMPLES_DIR`` / ``OUTPUT_DIR`` via a private ``_bot_core_dirs()``
  helper that late-imports them on each call.

Backward compat:

* ``bot.core`` re-exports every previously-public symbol — ``audit``,
  ``check_rate_limit``, ``RATE_LIMIT_PER_HOUR``, ``_rate_buckets``,
  ``_evict_lru_conversations``, ``sanitize_filename``, ``resolve_dest``,
  ``validate_path``, ``validate_input_path``, ``discover_file``,
  ``_build_trusted_dirs``, ``_ensure_trusted_dirs``,
  ``TRUSTED_DATA_DIRS``.
* External tests that monkey-patch ``server._core.TRUSTED_DATA_DIRS``
  (``tests/test_notebook_files.py``, ``tests/test_app_server.py``)
  continue to work — the symbol still resolves at the same dotted
  path.

Verification:

* Bot regression net (181 tests across ``test_bot_*``,
  ``test_oauth_regressions``, ``test_skill_listing``,
  ``test_routing_regression``, ``test_auto_routing_disambiguation``,
  ``test_app_server``, ``test_notebook_files``, ``tests/bot/``):
  181 passed, 1 skipped, 4 xfailed — no regressions.
* ``tests/eval/`` + ``tests/test_knowhow.py`` — 121 passed, 17
  deselected. Eval baseline 14/15 against ``deepseek-v4-pro``
  unaffected.
* ``bot/core.py``: 5049 → 4878 LOC (−171). Cumulative drop from main:
  5160 → 4878 LOC (−282).

Closes #114.
